+++ test/com/ansorgit/plugins/bash/lang/parser/HereDocParsingTest.java
@@ -1,13 +1,10 @@
 /*
- * Copyright 2010 Joachim Ansorg, mail@ansorg-it.com
- * File: HereDocParsingTest.java, Class: HereDocParsingTest
- * Last modified: 2010-06-06
+ * Copyright (c) Joachim Ansorg, mail@ansorg-it.com
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -25,7 +22,7 @@ import org.junit.Test;
  * @author jansorg
  */
 public class HereDocParsingTest extends MockPsiTest {
-    private MockFunction hereDoc = new MockFunction() {
+    private final MockFunction hereDoc = new MockFunction() {
         @Override
         public boolean apply(BashPsiBuilder psi) {
             return Parsing.file.parseFile(psi);
@@ -179,5 +176,15 @@ public class HereDocParsingTest extends MockPsiTest {
                 WORD, LINE_FEED,
                 FI_KEYWORD
         );
+    }
+
+    @Test
+    public void testTrailingSemicolon() throws Exception {
+        //issue 474
+        mockTest(hereDoc, Lists.newArrayList("cat", "<<", "EOF", ";", "\n", "X", "EOF"),
+                WORD, HEREDOC_MARKER_TAG, HEREDOC_MARKER_START, SEMI, LINE_FEED, HEREDOC_CONTENT, HEREDOC_MARKER_END);
+
+        mockTest(hereDoc, Lists.newArrayList("cat", "<<", "EOF", "&", "\n", "X", "EOF"),
+                WORD, HEREDOC_MARKER_TAG, HEREDOC_MARKER_START, AMP, LINE_FEED, HEREDOC_CONTENT, HEREDOC_MARKER_END);
     }
 }